Move saved configurations and enhance global runs
